Replace glob internals with PathBuf to reduce liklihood of issues on windows (#32)
Summary:
In the process of cleaning up tests from the previous two diffs, I found the issue where combining paths on windows resulted in a path with multiple separators: `path\\prefix\\to/my/directory` (or something along those lines).

I think the issue is that the above example is a valid `PathBuf`, and rust knows how to access it, but when converting to strings equality doesn't work.

Since we're doing a bunch of path manipulation with these globs, it's probably worth just keeping them in `PathBuf` form to make sure we handle them correctly everywhere. This diff replaces the underlying data type we're storing them as with `PathBuf`s, and only converts them to strings when performing matching or running the glob.

This fixes the issues in tests when I run it on both windows and linux machines.


Differential Revision: D71640818