Attempt to match wildcard all directories when matching glob pattern (#127)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/127

Kind of similar to the previous issue, when checking if a file matches a glob pattern, we fail to match if a pattern does not end in a wildcard and isn't directly matching a file (e.g. `my/root/dir` will not match `my/root/dir/some/file.py`).

We can fix this issue by retrying the pattern match with a `**` appended to the pattern.

Reviewed By: Solumin

Differential Revision: D73400445