Update package.json
Introduced via: docusaurus-plugin-internaldocs-fb@1.19.1
DOMPurify@3.2.4


The vulnerability in DOMPurify could allow an attacker to inject a malicious script into the application if user input is not correctly sanitized. In particular, this issue manifests in the handling of certain HTML elements or attributes, which can be exploited through a crafted payload that bypasses the sanitization process.