Use first non-error exit code interpreter as system interpreter (#449)
Summary:
Pull Request resolved: https://github.com/facebook/pyrefly/pull/449

On windows, `which` might find a system installation that doesn't exist because windows inserts a shim by default.

We can get around this by trying to run `<executable> --version` and seeing if it returns a non-zero exit code. If we add this to the functionality around finding system interpreters, we can filter to the first successful Python executable.

Differential Revision: D76292427