Update python interpreter in config to make verbose logging clearer
Summary:
It's confusing that we say `python_interpreter: None` in the `DEBUG Config is:` output during verbose logging. It's confused me and several users, so let's fix that bug here so that we can stop being confusing.

p.s. I think I made this change a while ago and just forgot to write the found interpreter :(

Differential Revision: D76307599