Fix non-determinism in autocomplete integration test
Summary:
The non-determinism is likely coming from that there are multiple matches, and one match is on the edge of being either matched or not matched.

This diff addresses the issue by running the test on an extremely long name instead, so that the fuzzy matcher will always match it with high confidence.

Differential Revision: D77039902