Fix bugs in attribute narrowing
Summary:
fixes https://github.com/facebook/pyrefly/issues/547

BUG 1: In the linked issue, we observed that the narrowing for an attribute was invalidated by assigning to a subscript of that attribute. This is because we accidentally check the narrowing information for `next_facet` instead of `facet` in `NarrowedFacets::update_for_assignment`, causing the program to take a different branch and overwriting the existing narrow.

BUG 2: Once I fixed the first bug, the attribute narrowing was no longer being wiped, but the subscript narrowing did not get applied. This is because the new branch the program takes hit another bug in `NarrowedFacet::update_for_assignment`, which stops execution if the current narrowed facet is a leaf (even if we want to narrow a subtree). I ended up inlining the fixed logic directly since the function is only called once.

Differential Revision: D77175069