Fix up test_find_module_prefixes
Summary:
This code finds module prefixes in the order they appear on disk. Often files written in one order will read in that order, but there is no guarantee, and it's file system dependent. To make sure these tests are deterministic on all file systems, sort the result.

Note that before the result is used in LSP we sort it anyway, so it doesn't need sorting in the real method.

Differential Revision: D77373942