Use `AnswersSolver::new` from `Answers`
Summary:
The `answers.rs` module is getting too busy with cycle breaking work
in progress. I was going to put off factoring things out becasue it's
a bit bikesheddy, but I realized that I'll probably be slower actually
implementing things if I don't isolate the tricky stuff better.

The easiest way to break up `answers` is to move the AnswerSolver out;
the main blocker for that is the use of direct struct construction.

Reviewed By: ndmitchell

Differential Revision: D76948126

fbshipit-source-id: dbfd922cccdac89da3bc7d4c31d3651050d42ccf