Create plumbing for cycle breaking
Summary:
This diff doesn't actually do anything, but I wire up the plumbing
for `AnswersSolver` to be aware of cycle-breaking.

The idea here is that solving cyclic graphs involves breaking cycles,
and the cycles form a stack because you can encounter a cycle while
solving a different cycle. Breaking cycles deterministically will
require doing more than we do to day, and actually tracking some
data as we resolve the cycle - we'll use the `Cycle` data structure,
which at the moment is just a unit type, for that.

Cycles are thread-local, so the cycles get passed around and created
in exactly the same places as the calculation stack; we might want to
eventually coalesce these into some kind of "thread state" data structure,
but I'd rather delay that kind of bikeshedding until after the core
algorithm is working.

Reviewed By: ndmitchell

Differential Revision: D76948125

fbshipit-source-id: 55bd1b10ef4bc9e6ea7ab78b9f4d7c80e69ef362