Simplify `get_idx` a bit
Summary:
A lot of the complexity in the existing logic came from making every branch of
a match return a type complicated enough to control another match because we were
trying to calculate or fetch the result and then record recursive correspondence
independently.

But these two operations aren't actually independent - the only scenario where we would
record recursive is when we've just stored an answer. Inlining that makes things
a bit simpler, the types make more sense and we get rid of one of the blocks in
`get_idx`.

Reviewed By: ndmitchell

Differential Revision: D76948142

fbshipit-source-id: 34fe1a0082805a151ed0adec5035bc79a946a6d0