Factor some things out of `get_idx`
Summary:
In order to implement deterministic cycle breaks, I'm going to wind up needing
a double match, where the exact logic will depend on some combination of the
cycle state of the current idx and the calculation state.

It's a little hard to see how to do that well without rearranging some of the
fiddly logic more clearly - this diff mostly leaves things as they were but
moves the fiddling into helpers.

Reviewed By: ndmitchell

Differential Revision: D76948134

fbshipit-source-id: f533c3323a2724030385e1fa1084003f85fd07a1