Validate subtype relation for TypeIs
(This diff depends on #656)

According to the spec for TypeIs
https://typing.python.org/en/latest/spec/narrowing.html#typeis,
the return type in `TypeIs` must be a subtype of the input type being
narrowed. This diff adds a check to that effect for the required
subtyping relation, and adds appropriate test cases.

After this change, all the conformance tests for typeis should pass.