Raise error on variable use before write
Closes #15, #248

This change adds additional checks when building bindings, so that
variables that are used before initialised are correctly flagged as
errors.

Currently, we error on uninitialised variables by checking their flow
style. If the flow style is uninitialised or partially uninitialised, an
error is generated. When no flow style information is obtainable (which
is the case for use before write), a default "other" flow style value is
returned.

In this diff, we perform additional checks before returning a default
"other" value. By looking up the scopes for static information for
variables, we can correctly detect if a variable is declared in an upper
scope or in the current scope, and return "uninitialised" flow style
when the variable is used before write.

When doing so, we need to consider the special case for type usages,
so that name resolutions in types do not fail.