Fix caught exception leaking through `except` scope
As reported in #154, if a caught exception is bound to a name `e` in a
`try`/`except` block, the caught name `e` should be out of scope after
the current `except` scope.

This is reflected in the spec as follows
https://docs.python.org/3/reference/compound_stmts.html#except-clause:

> When an exception has been assigned using as target, it is cleared at the end of the except clause.

This PR aims to model this effect by marking the caught exception name
as uninitialised at the end of the `except` clause.