Remove gemfile
Summary:
Just seeing what happens if I remove this.

It was copied over from flow's website and has not been changed since, and I am confused
about why it's here. The docs make no mention of any ruby tooling.

If CI passes with this removed, I'm almost certain it is not needed (the website built fine without it)

Differential Revision: D78703559