Try turning on rust-cache for the `pyrefly` workflow
Summary:
I'm not totally satisfied that I understand perfectly how this works, it's
at least possible that our override of `CARGO_HOME` will impact it on
windows.

But it's worth trying. The exact config snippet I pulled from a buck
example.

I think the pyrefly workflow accounts for the majority of our rust build
time, so I think that if this works, we should see some actual results
in both turnaround time and compute use.

Differential Revision: D78710986