sort before deduping (#723)
Summary:

In the IDE, I almost always see duplicate completions for stdlib functions. This is because we add all search paths and only dedup results by neighbors. We end up finding the builtin in-memory typeshed, the one in the interpreter, and any others. instead, we should sort before deduping to remove all duplicates. Since VSCode ranks our results anyway, this should have no side effects in the IDE.

Note: the test added in this diff fails before this diff (there are duplicate MutableMapping), but i'm unable to commit it without this change because auto-import is not deterministic in which module it adds first. This sorting also helps us write similar tests in the future.

Reviewed By: ndmitchell

Differential Revision: D78700243