Warn on importing deprecated symbols (tentative implementation)
This commit imports warning for importing deprecated symbols.

We collect deprecated symbols when building the exports, by checking the
decorators when processing function definitions and class definitions.
The deprecated symbols are then stored as part of the export
information.

When importing new symbols, we check whether the exported name is
deprecated and generate a warning accordingly.