Make find_one_part return remaining search roots
Summary:
# This diff...

Updates `find_one_part` to return the rest of the search roots, so that we can continue searching for an import if we want to attempt to find a better import alternative.

# This diff stack...

implements preference for `.pyi` files over `.py` files in import finding, even if the import occurs later in the includes list. For a directory that looks like this:
```
root/
|- a
|  |- c.py
|
|- b
   |- c.pyi
```
and a site package path or search path of `[root/a, root/b]`, we would locate `import c` as `c.pyi`.

## Implementation

So far, we've:
- pulled out the functionality for `find_one_part` in a single root (`find_one_part_in_root`)
- created `.pyi` and `.py` `FindResult` single file variants
- removed `use_untyped_imports` logic

Next, we will:
- create a comparator for `FindResult` variants to select the best option
-

## Approaches (if you're interested :) )

There are two approaches I was considering for this:
1. Keep the implementation mostly the same, but have some boolean flag or something that tells us if we can search for `.py` files. Then we just perform the search twice.
    1. The advantage of this approach is mostly keeping things that we trust are working right now mostly the same
2. Change the implementation to check all `RegularPackage`s, instead of just taking the first one found, taking `.pyi` files over `.py` files if any are found. At the end, only return the first `RegularPackage`, if there is one, or the first `.pyi` file if one is found.
    1. The advantage of this approach is that we don't crawl the filesystem multiple times, rechecking everything, with the only difference being checking for the presence of a `.py` file.
    2. I believe this should hold mostly the same semantics in the end, since we'll still prioritize `RegularPackage` results over anything else, and we preserve the order we find them in, returning the first one at the end of the search if `RegularPackage` is the final resulting type in `continue_find_module`. The difference is that if there happen to be two `RegularPackage`s, we *will* end up looking at the second one for a valid import if we can't find it in the first, which we wouldn't do previously.

I decided to implement option two, since it seems like a better approach in terms of efficiency and lets us extend the searching behavior in the future if we decide we want to do that.

Differential Revision: D79490956