feat(code-action): add ignore comment to suppress error
This PR address #675.

It works by
1. By gathering all the errors whose range contains the current
   cursor.
2. Finding the column position of the first non-whitespace character of
   the line the cursor is on. Call this `column offset`.
3. Inserting an ignore comment on the line above the cursor; the
   comment is preceded by `column offset` single-spaces to match the
   prior indentation.

In my tests I have found this simple scheme to be surprisingly robust.
My first attempt used the ast exclusively, but the ast does not contain
comments, nor a way to insert _comments_ into the tree, meaning, just for this
one use case of programmatically adding ignore comments, I cannot
update the document using the ast, and I found myself having to
calculate whitespace offsets in my first ast only attempt.

It is also worth pointing out that some LSPs might not _want_ to provide
this kind of code action--they want to encourage the user to address the
underlying cause, not ignore it. But because Pyrefly is new, there are
many false positives and so I think this will be a welcome addition.
Plus, the suppression comments in this code action reference
specific error codes, so the user in the future can address the lints
precisely.

Future work is needed to address different types of ignore comments and
a code action to

- [ ] Add all ignore comments to a module
- [ ] Add a top-level ignore comment to a module

Both should be relatively straightforward to do by using the utility
functions I added in `pyrefly/lib/state/ide.rs`

Let me know your thoughts on this.

Thanks!