[RFC] Warn if deprecated function is referred
This diff implements the last missing feature for deprecation warnings:
to warn when the deprecated function is referred (e.g. passed as
function reference).

The implementation checks when typechecking an expression, whether the
type of an expression is a deprecated function/bound method. If that's
the case, a deprecated warning is generated.

If we only add this check, we might run into the issue that a deprecated
function call would generate 2 deprecation warnings, one when the
function is typechecked, and another when the call is typechecked.

To avoid the duplication, we avoid emitting the deprecation warning when
typechecking a function call, instead we emit the deprecation warning
when generating the call target.

As a consequence, the error message of deprecation warning is changed
from "Calling" a deprecated function to "Reference to" a deprecated
function. Suggestions are welcome for how to improve clarify of this
error message.