Handle scoping correctly for class fields in body
Fixes #264.

This diff addresses the scoping issues of class fields in nested scopes.

When a class field is defined in a class body, the scope of the field
should only be available in the immediate class block, unless for
annotations. This is reflected in the Python doc in
https://docs.python.org/3/reference/executionmodel.html#resolution-of-names

To address this issu, we check the flow style when looking up a
variable. If the variable is defined as a class field, we apply
appropriate checks to avoid the incorrect scoping rule from being
applied.