Add config so definition typeshed tests work without interpreter installed
Summary:
A bug report came in that Pyrefly doesn't build in `nix` when there's no Python interpreter available. Turns out we need it for some go-to-def functionality testing, but it isn't actually needed for things to work/Pyrefly to build.

[Original report](https://discord.com/channels/1319086885024567347/1319086885024567350/1414470301470953573)

This diff adds a config and moves some files around to get it working without an interpreter.

Differential Revision: D81943027