Support enums inheriting from multiple classes
Summary:
This decreases  django's test_enum's false positives.
 52 errors -> 44 errors



I am not familiar with the enum code yet so would love some feedback here. 

Currently, we don't support iterating over an enum when it inherits from multiple classes. This is a bug in normal enums and an important one to address in django enums. 

I have a workaround, which is to store the enum iterable type in the enum metadata and then special case it in the solver.

Can I get some initial thoughts about this as a temporary workaround or if we need to hold off on this and explore other options?

Differential Revision: D83203854