Handle integer choice downstream (#1154)
Summary:

We do so by extracting the first element of the type when we see a django integer choice enum.

I also noticed that we use the basetype, so I promoted the type as well to match runtime.

We make those changes in ClassField where we create those fields, and the place where this will be reflected is in the attr code.

This decreased the errors by 7 in the enums tests, so we are at 52

Reviewed By: rchen152

Differential Revision: D83099105