Check for compatibility when inherting from multiple classes
This diff is an attempt to address the issue in #943.

When a class inherits from multiple base classes, we check whether the
intersection of the types of the fields is never. If that's the case, we
raise an error.

Note 1:
There are some errors when type checking the builtins (which is
reflected by errors in scrut tests). They seem to be caused by
subtyping checks of self types in functions and generic function types.

Note 2:
Should this be a new error category? I'm currently reusing the closest
error category for inconsistent overloads, but that looks very specific.