collect type params at bindings phase for legacy explicit type alias
Summary:
fixes https://github.com/facebook/pyrefly/issues/1162

This is related to our (problematic) handling of legacy type aliases as well as our internal representation of type aliases.

For generic legacy type aliases, we traverse the resolved type of the RHS to determine the type parameters. However, when the RHS is a specialization of another type alias, the type we see is the aliased type because `untype()` eagerly resolves aliases. The order of the type arguments during specialization may not match the order that they appear in the aliased type, resulting in this bug.

https://github.com/facebook/pyrefly/issues/1162#issuecomment-3353329339 explains a bit more.

The solution in this diff is to use the legacy type parameter collector (also used in function and class definitions) to collect type parameters at the bindings phase when we know for sure we're dealing with a type alias (the assignment is annotated with TypeAlias or the RHS is a UnionType, for example). If we're not sure, then we fall back to the previous behavior of collecting type params from the resolved type.

One long term fix is to have a type that represents a specialized type alias, like we do for classes using`ClassType`.

Differential Revision: D83597568