differentiate between notebook and source files in language server
Summary:
Since DidChange/DidOpen doesn't send the entire notebook JSON over, we have to be able to construct a notebook struct within the language server instead of relying on the file system loader.

If the language server provides a notebook struct, we'll use that instead of loading it from disk.

Huge amount of plumbing to get this all the way through, but I think this should work.

Differential Revision: D85809461