Handle the case for modifying a final field as class var
Following up from previous commit, we also handle the case we modify a
final field in dataclass of the class instead of an instance.

To do so, we need to pass on the class base in a helper function,
alongside the instance class. Maybe we should use an Either type since
the two should not be passed at the same time.