preserve notebook contents in didChange
Summary:
There was a bug in this impl, where didChange would actually set all cell contents to empty string, except for any cells that were added in the change request.

This diff preserves existing cell contents, and also refactors the code to avoid unnecessarily turning a map into a vec and back to a map.

Differential Revision: D86064878