add optional display name to Type::Union
Summary:
This is an alternative to D84849556, which aims to solve https://github.com/facebook/pyrefly/issues/1219

Instead of introducing a new type variant, this diff introduces a new field in Type::Union to hold an optional display name. It will be populated when resolving a type alias.

Compared to the other approach:

benefits
- less plumbing
- type checking behavior unaffected

drawbacks
- only applies to type aliases of unions (though arguably this is where readability matters most)
- name information is not preserved during union flattening (so if we union a type alias with something else, the result is the full expanded type)

Differential Revision: D87558541