Combine the top to matches of `calculate_class_field`
Summary:
I'm finally ready to combine the first two match statements of
`calculate_class_field`, eliminating the `value` variable and
`value_storage` in the process because we can just compute the
`value_ty` in one shot.

This is a major step along the path to being able to represent
the cases of class field more clearly - even thought the new match
is a bit more complex, it can now be understood by reading it (in linear
time) - we no longer have to understand the interactions between
two matches (which is in some sense quadratic) to understand what `value`
will be for any given case.

Reviewed By: grievejia

Differential Revision: D87653413

fbshipit-source-id: 272e6bff7345d832f10e82df19135b7c6f6d7023