Inline get_class_field_initialization
Summary:
We want `RawClassFieldInitialization` and `initial_value` to go away;
these aren't necessary, they are just tech debt accumulated.

As a result, we don't want to have a helper function, because that makes
it harder to deduplicate and simplify the match statements. In preparation
for coming simplifications, let's just inline.

This makes things messier in the short term but will pay off as we keep
cleaning things up.

Reviewed By: grievejia

Differential Revision: D87653407

fbshipit-source-id: 79b14889b71534b70bd591f89eca64939a2f7b28