Just pass down inferred_from_method as a boolean flag
Summary:
We want to get rid of RawClassFieldInitialization - it's unnecessary once
we've refactored the code to make more sense. As a result, it's important to
avoid depending on it where possible; in this case, let's keep the match in
`calculate_class_field` (which is where we need to refactor) and just pass
a flag down.

Reviewed By: rchen152

Differential Revision: D87653414

fbshipit-source-id: 030b4e353d867f6cd9c5696bc5f80ee9610b0f93