preserve quantified information when unpacking quantified type var tuple
Summary:
Unpacking a `tuple[*Ts]` hits some fallback code and turns the contents into `object`

We fix this by introducing an `ElementOf[Ts]` type, which is similar to `Type::QuantifiedValue` for TypeVar but it represents an element of the TypeVarTuple, instead of the whole thing.

So when we try to iterate some `tuple[*Ts]`, we produce an unspecified number of `ElementOf[Ts]`, and when we try to make an unbounded tuple with `ElementOf[Ts]` contents we actually create `tuple[*Ts]`.

There are additional guards against unpacking the same type var tuple twice in a tuple expression or function call, since the lengths will not match up unless the typevartuple resolves to empty tuple.

fixes https://github.com/facebook/pyrefly/issues/1268

Reviewed By: stroxler

Differential Revision: D87694394

fbshipit-source-id: 19be3f7282d6e4dfec13e0df5523e2ade658e63a