Add OutputWithLocations struct.
Summary:
This stack introduces a new struct called OutputWithLocations that will be used along with fmt_helper_generic in order to make certain parts of inlay hints clickable.

The idea here is that now instead of writing parts of the type directly to a formatter, there is a generic TypeOutput that will be taken in can handle the collection of parts itself. In this case, we will be using an OutputWithLocations in order to collect the parts of the type and also get the location of the types definition if it is relevant.

The rest of this stack will handle the actual implementation of this logic.

There are a few things that currently are known limitations that will be added as follow ups:
1. Currently tuple types will not be clickable. For example if you have `tuple[int]`, the `int` portion of the type will be clickable but the `tuple` portion will not be. A test has been added that covers this case.
2. Types coming from typing.py will not be clickable. For example, the type `Literal[1]` will not be clickable. This is something that will be addressed in a followup. A test has been added which covers this case.

Introduces a new OutputWithLocations struct that implements TypeOutput to capture formatted type strings along with their source code locations. This struct collects type parts as a vector of (String, Option<TextRangeWithModule>) tuples, enabling clickable type references in IDE features. Also exposes fmt_helper_generic as public to support the new location-aware formatting infrastructure.
Adds the OutputWithLocations struct that implements the TypeOutput trait. This struct stores type parts as a vector of (String, Option<TextRangeWithModule>) tuples, allowing us to capture both the formatted type string and its source location. The struct provides a new() constructor that takes a TypeDisplayContext and a parts() getter to access the collected parts. This is the first step in enabling clickable type references in IDE features.

Reviewed By: ndmitchell

Differential Revision: D87708533

fbshipit-source-id: 09d9c705974ff373a291d43880ffa752c198825b