Add test showing that tuple type does not have a location.
Summary:
Currently if you have a type like `tuple[int]`, the int portion of the type will have a location but the `tuple` portion does not. This diff adds a test to cover this case and will be updated once that functionality is implemented.

Also updates the formatting for tuples so that `tuple` and `[` are separate parts rather than just a single `tuple[` part.

Reviewed By: kinto0

Differential Revision: D87642082

fbshipit-source-id: dcd0e44785a87ba608ef2431fb0c8a29cae0a816