Support `object.__new__(C)`
Summary:
We have a pair of bug reports noting that the `tqdm`
library is essentially unsupported by Pyrefly, because
the constructor uses `object.__new__(C)` as part of the
constructor flow for a class and this causes Pyrefly
to infer the class as being of type `object`.

This is incorrect, `object.__new__` has type
`(self: type[Self]) -> Self`, and since `Self` behaves
as a type var with upper bound `object` here; this
means if we pass any class `C` to it we ought to get
back an instance of `C`.

This commit fixes the logic for just `object.__new__`,
other unbound method calls are still incorrect according
to the spec because a more general "`Self` is a type var"
fix will be more complicated and I want a quick fix for the
fact that Pyrefly can't model the `tqdm` library at all right now.

Should fix
https://github.com/facebook/pyrefly/issues/1856
and
https://github.com/facebook/pyrefly/issues/1375

Reviewed By: yangdanny97

Differential Revision: D89479870

fbshipit-source-id: 936e2b362296afcc5e94e0961ecfac40484dde60