Specialize HeavyTaskQueue to HeavyTask (again)
Summary:
I was planning to defunctionalize these queues, and I was going to incrementally replace HeavyTask with data. But then I found a shortcut (passing &Server as an argument).

Defunctionalization would still be a good idea, but right now the more important thing is to add in telemetry. Specializing the queue makes this easier.

Reviewed By: connernilsen

Differential Revision: D89595800

fbshipit-source-id: bd9b9490d7569cafc365b1f46b8bcf2e5544c01b