Return whether ensure inserted a value
Summary:
I want to add a counter here, but since we can have two threads race to insert a value, I want to observe whether we succeeded or not.

The caller in get_module_ex already had similar code to see if we also created the module for the first time. I slightly modified this to be simpler, by checking both `inserted` and `created`.

Reviewed By: connernilsen

Differential Revision: D89599952

fbshipit-source-id: 1385c283a147437f2a4509dc6993d9d2b0a98cb6