Forbid re-assignment of `Final` fields in constructor
Previously, the check for `Final` fields in classes did not check
whether the final fields are provided with an initial value. This diff
adds check for whether the final fields have been initialised in their
declaration, and raises an error if they are re-assigned in constructors
or other methods that are permitted to modify attributes.

Fixes one failing test in the conformance testsuite.